Move all *.spec files to test/ dir
