+++ test/wait.spec.ts
@@ -1,5 +1,5 @@
-import wait from "./wait";
-import { time, WAIT_TIME, TIMEOUT_MARGIN } from "./time-test-helpers";
+import wait from "../src/wait";
+import { time, WAIT_TIME, TIMEOUT_MARGIN } from "./helpers/time";
 
 it("resolves after the specified amount of time", async () => {
 	const finish = time() + time(WAIT_TIME);
