+++ tsconfig.json
@@ -13,7 +13,7 @@
     "forceConsistentCasingInFileNames": true,
   },
   "include": [
-    "src"
+    "src",
   ],
   "exclude": [
     "**/*.spec.*"
